chore: move implementation to new plugin
